backend/requirements.txt
django==4.2.11
djangorestframework==3.15.1
django-cors-headers==4.3.1
Pillow==10.0.0
numpy==1.26.4
transformers==4.38.2
langchain==0.1.13 
langchain-huggingface==0.0.3
python-decouple==3.8
gunicorn==21.2.0
requests==2.31.0
django-allauth==0.61.1
scikit-image==0.22.0
opencv-python==4.9.0.80
matplotlib==3.8.3